🛠️ Restrict dependency resolution to Python < 3.14 (#29)
## Description

This PR adjusts the `uv` resolution strategy so that, hopefully,
renovate won't fail the next time it tries to update the lockfile.
Catalyst ships ABI3 wheels, but some of its dependencies don't support
Python 3.14 yet. This seems to create problems when trying to resolve
the lockfile from Python 3.14.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

Signed-off-by: burgholzer <burgholzer@me.com>